changes to updater, protos, and construction
- Updater will not use hard coded user id and instead rely on a sorted
set in redis named "user_updates"
- Protobufs changed to use int32 for everything but timestamps since
they are formatted as numbers in JSON (protojson) (int64s are too big to
be represented as JS/JSON numbers)
- First construction of each building is free
- Construction times are appended (previously constructions were
isolated, as if the mice would work in parallel)

+++ internal/mapping.go
@@ -16,18 +16,18 @@ func (gs *GameState) ToStateChangeMessage() *ServerMessage {
 	pop := &GameStatePatch_PopulationPatch{}
 
 	if gs.Population != nil {
-		pop.Uneducated = &wrapperspb.Int64Value{ Value: gs.Population.Uneducated }
-		pop.Chefs =      &wrapperspb.Int64Value{ Value: gs.Population.Chefs }
-		pop.Salesmice =  &wrapperspb.Int64Value{ Value: gs.Population.Salesmice }
-		pop.Guards =     &wrapperspb.Int64Value{ Value: gs.Population.Guards }
-		pop.Thieves =    &wrapperspb.Int64Value{ Value: gs.Population.Thieves }
+		pop.Uneducated = &wrapperspb.Int32Value{ Value: gs.Population.Uneducated }
+		pop.Chefs =      &wrapperspb.Int32Value{ Value: gs.Population.Chefs }
+		pop.Salesmice =  &wrapperspb.Int32Value{ Value: gs.Population.Salesmice }
+		pop.Guards =     &wrapperspb.Int32Value{ Value: gs.Population.Guards }
+		pop.Thieves =    &wrapperspb.Int32Value{ Value: gs.Population.Thieves }
 	}
 
 	p := &GameStatePatch{
 		Lots: lotsPatch,
 		Resources: &GameStatePatch_ResourcesPatch{
-			Coins: &wrapperspb.Int64Value{ Value: gs.Resources.Coins },
-			Pizzas: &wrapperspb.Int64Value{ Value: gs.Resources.Pizzas },
+			Coins: &wrapperspb.Int32Value{ Value: gs.Resources.Coins },
+			Pizzas: &wrapperspb.Int32Value{ Value: gs.Resources.Pizzas },
 		},
 		Population: pop,
 		TrainingQueue: gs.TrainingQueue,
